Reformulate everything for the new traits
