+++ src/fish.rs
@@ -1,6 +1,6 @@
 #![cfg(feature = "fish")]
 
-use crate::{ascii::Char, sealed, util::u8_to_hex_escape_uppercase_x, Quotable, Quoter};
+use crate::{ascii::Char, util::u8_to_hex_escape_uppercase_x, Quotable, QuoteInto};
 
 /// Quote byte strings for use with fish.
 ///
@@ -23,17 +23,33 @@ use crate::{ascii::Char, sealed, util::u8_to_hex_escape_uppercase_x, Quotable, Q
 #[derive(Debug, Clone, Copy)]
 pub struct Fish;
 
-impl Quoter for Fish {}
-
-impl sealed::Quote for Fish {
-    fn quote<'a, S: ?Sized + Into<Quotable<'a>>>(s: S) -> Vec<u8> {
-        Self::quote(s)
+impl QuoteInto<Vec<u8>> for Fish {
+    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut Vec<u8>) {
+        Self::quote_into(s, out);
     }
 }
 
-impl sealed::QuoteInto for Fish {
-    fn quote_into<'a, S: ?Sized + Into<Quotable<'a>>>(s: S, sout: &mut Vec<u8>) {
-        Self::quote_into(s, sout)
+impl QuoteInto<String> for Fish {
+    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut String) {
+        Self::quote_into(s, unsafe { out.as_mut_vec() })
+    }
+}
+
+#[cfg(unix)]
+impl QuoteInto<std::ffi::OsString> for Fish {
+    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut std::ffi::OsString) {
+        use std::os::unix::ffi::OsStringExt;
+        let s = Self::quote(s);
+        let s = std::ffi::OsString::from_vec(s);
+        out.push(s);
+    }
+}
+
+#[cfg(feature = "bstr")]
+impl QuoteInto<bstr::BString> for Fish {
+    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut bstr::BString) {
+        let s = Self::quote(s);
+        out.extend(s);
     }
 }
 
@@ -50,7 +66,7 @@ impl Fish {
     /// # Examples
     ///
     /// ```
-    /// # use shell_quote::{Fish, Quoter};
+    /// # use shell_quote::{Fish, Quote};
     /// assert_eq!(Fish::quote("foobar"), b"foobar");
     /// assert_eq!(Fish::quote("foo 'bar"), b"foo' \\'bar'");
     /// ```
@@ -74,7 +90,7 @@ impl Fish {
     /// # Examples
     ///
     /// ```
-    /// # use shell_quote::{Fish, Quoter};
+    /// # use shell_quote::{Fish, Quote};
     /// let mut buf = Vec::with_capacity(128);
     /// Fish::quote_into("foobar", &mut buf);
     /// buf.push(b' ');  // Add a space.
