+++ src/lib.rs
@@ -15,7 +15,7 @@ mod ascii;
 mod bash;
 mod fish;
 mod sh;
-pub(crate) mod util;
+mod util;
 
 #[cfg(feature = "bash")]
 pub use bash::Bash;
@@ -33,6 +33,8 @@ pub type Dash = sh::Sh;
 /// for [`Bash`].
 #[cfg(feature = "bash")]
 pub type Zsh = bash::Bash;
+
+// ----------------------------------------------------------------------------
 
 /// Extension trait for pushing shell quoted byte slices, e.g. `&[u8]`, [`&str`]
 /// – anything that's [`Quotable`] – into byte container types like [`Vec<u8>`],
@@ -60,16 +62,25 @@ impl<T: ?Sized> QuoteExt for T {
 /// e.g. `&[u8]`, [`&str`] – anything that's [`Quotable`] – into owned types
 /// like [`Vec<u8>`], [`String`], [`OsString`] on Unix, and [`bstr::BString`] if
 /// it's enabled.
-pub trait QuoteRefExt<Output> {
-    fn quoted<Q: Quoter>(self, q: Q) -> Output;
+pub trait QuoteRefExt<Output: Default> {
+    fn quoted<Q: Quote<Output>>(self, q: Q) -> Output;
 }
 
 impl<'a, S> QuoteRefExt<Vec<u8>> for S
 where
     S: ?Sized + Into<Quotable<'a>>,
 {
-    fn quoted<Q: Quoter>(self, _q: Q) -> Vec<u8> {
-        Q::quote(self)
+    fn quoted<Q: Quote<Vec<u8>>>(self, _q: Q) -> Vec<u8> {
+        Q::x_quote(self)
+    }
+}
+
+impl<'a, S> QuoteRefExt<String> for S
+where
+    S: ?Sized + Into<Quotable<'a>>,
+{
+    fn quoted<Q: Quote<String>>(self, _q: Q) -> String {
+        Q::x_quote(self)
     }
 }
 
@@ -78,10 +89,8 @@ impl<'a, S> QuoteRefExt<OsString> for S
 where
     S: ?Sized + Into<Quotable<'a>>,
 {
-    fn quoted<Q: Quoter>(self, _q: Q) -> OsString {
-        use std::os::unix::ffi::OsStringExt;
-        let quoted = Q::quote(self);
-        OsString::from_vec(quoted)
+    fn quoted<Q: Quote<OsString>>(self, _q: Q) -> OsString {
+        Q::x_quote(self)
     }
 }
 
@@ -90,129 +99,22 @@ impl<'a, S> QuoteRefExt<bstr::BString> for S
 where
     S: ?Sized + Into<Quotable<'a>>,
 {
-    fn quoted<Q: Quoter>(self, _q: Q) -> bstr::BString {
-        let quoted = Q::quote(self);
-        bstr::BString::from(quoted)
+    fn quoted<Q: Quote<bstr::BString>>(self, _q: Q) -> bstr::BString {
+        Q::x_quote(self)
     }
 }
 
 // ----------------------------------------------------------------------------
 
-pub(crate) mod sealed {
-    pub trait Quote {
-        /// Quote/escape a string of bytes into a new [`Vec<u8>`].
-        fn quote<'a, S: ?Sized + Into<super::Quotable<'a>>>(s: S) -> Vec<u8>;
-    }
-
-    pub trait QuoteInto {
-        /// Quote/escape a string of bytes into an existing [`Vec<u8>`].
-        fn quote_into<'a, S: ?Sized + Into<super::Quotable<'a>>>(s: S, sout: &mut Vec<u8>);
-    }
-}
-
-/// A trait for quoting/escaping a string of bytes into a shell-safe form.
-///
-/// **This trait is sealed** and cannot be implemented outside of this crate.
-/// This is because the [`QuoteExt`] implementation for [`String`] must be sure
-/// that quoted bytes are valid UTF-8, and that is only possible if the
-/// implementation is known and tested in advance.
-pub trait Quoter: sealed::Quote + sealed::QuoteInto {}
-
-// ----------------------------------------------------------------------------
-// -- New traits --------------------------------------------------------------
-// ----------------------------------------------------------------------------
-
-/// REPLACES [`quoter::QuoterSealed::quote_into`].
+/// Quoting/escaping a string of bytes into a shell-safe form.
 pub trait QuoteInto<OUT: ?Sized> {
+    /// Quote/escape a string of bytes into an existing container.
     fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut OUT);
 }
 
-impl QuoteInto<Vec<u8>> for Bash {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut Vec<u8>) {
-        Self::quote_into(s, out);
-    }
-}
-
-impl QuoteInto<String> for Bash {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut String) {
-        Self::quote_into(s, unsafe { out.as_mut_vec() })
-    }
-}
-
-#[cfg(unix)]
-impl QuoteInto<OsString> for Bash {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut OsString) {
-        use std::os::unix::ffi::OsStringExt;
-        let s = Self::quote(s);
-        let s = OsString::from_vec(s);
-        out.push(s);
-    }
-}
-
-#[cfg(feature = "bstr")]
-impl QuoteInto<bstr::BString> for Bash {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut bstr::BString) {
-        let s = Self::quote(s);
-        out.extend(s);
-    }
-}
-
-impl QuoteInto<Vec<u8>> for Fish {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut Vec<u8>) {
-        Self::quote_into(s, out);
-    }
-}
-
-impl QuoteInto<String> for Fish {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut String) {
-        Self::quote_into(s, unsafe { out.as_mut_vec() })
-    }
-}
-
-#[cfg(unix)]
-impl QuoteInto<OsString> for Fish {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut OsString) {
-        use std::os::unix::ffi::OsStringExt;
-        let s = Self::quote(s);
-        let s = OsString::from_vec(s);
-        out.push(s);
-    }
-}
-
-#[cfg(feature = "bstr")]
-impl QuoteInto<bstr::BString> for Fish {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut bstr::BString) {
-        let s = Self::quote(s);
-        out.extend(s);
-    }
-}
-
-impl QuoteInto<Vec<u8>> for Sh {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut Vec<u8>) {
-        Self::quote_into(s, out);
-    }
-}
-
-#[cfg(unix)]
-impl QuoteInto<OsString> for Sh {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut OsString) {
-        use std::os::unix::ffi::OsStringExt;
-        let s = Self::quote(s);
-        let s = OsString::from_vec(s);
-        out.push(s);
-    }
-}
-
-#[cfg(feature = "bstr")]
-impl QuoteInto<bstr::BString> for Sh {
-    fn x_quote_into<'q, S: ?Sized + Into<Quotable<'q>>>(s: S, out: &mut bstr::BString) {
-        let s = Self::quote(s);
-        out.extend(s);
-    }
-}
-
-/// REPLACES [`quoter::QuoterSealed::quote`].
+/// Quoting/escaping a string of bytes into a shell-safe form.
 pub trait Quote<OUT: Default>: QuoteInto<OUT> {
+    /// Quote/escape a string of bytes into a new container.
     fn x_quote<'q, S: ?Sized + Into<Quotable<'q>>>(s: S) -> OUT {
         let mut out = OUT::default();
         Self::x_quote_into(s, &mut out);
@@ -220,20 +122,8 @@ pub trait Quote<OUT: Default>: QuoteInto<OUT> {
     }
 }
 
-/// Blanket impl for anything that already has a [`QuoteInto`] impl.
+/// Blanket [`Quote`] impl for anything that has a [`QuoteInto`] impl.
 impl<T: QuoteInto<OUT>, OUT: Default> Quote<OUT> for T {}
-
-fn _foo1() {
-    let mut a: OsString = Bash::x_quote(b"123");
-    Bash::x_quote_into("foo", &mut a);
-}
-
-fn _foo2() {
-    let mut string: String = Bash::x_quote(b"123");
-    string.push_quoted(Bash, "bar");
-    let mut os_string: OsString = Fish::x_quote("xyz");
-    os_string.push_quoted(Fish, "wobble");
-}
 
 // ----------------------------------------------------------------------------
